Version 1.0.4
removed rlang and prettycode Imports.

Created data/ directory and moved hurricanes.Rdata from Chapter 6 to there. Uncertain if I am happy with this development. It may be better to wait until we get to data frames and lists to include data. We may want to put other data in their like the ranked choice voting data.

Experimenting with attaching a new environment in the search path and parking persistent objects there.

Need to create a mechanism where by objects are created at a chapter level. run_tut() now detaches this holding area when it exits.

setup.R is now responsible for the creation of the holding environment and parking objects in there.